♻️ Migrate the schemas into the new design sub-package *core.database*.
