+++ gearmeshing_ai/core/database/schemas/__init__.py
@@ -1,0 +1,14 @@
+"""
+Database schema models for API requests and responses.
+
+This package contains Pydantic-based schema models for API serialization/deserialization.
+These schemas are separate from entity models to allow independent evolution of API contracts
+and database representations.
+"""
+
+from . import agent_configs, chat_sessions
+
+__all__ = [
+    "agent_configs",
+    "chat_sessions",
+]
