--- gearmeshing_ai/server/models/__init__.py
@@ -1,6 +1,0 @@
-"""
-Database Models.
-
-This module exports domain-specific models for agent configuration and chat sessions.
-For ORM models used by repositories, see gearmeshing_ai.agent_core.repos.models.
-"""
